Fix: matching wrong if cursor on open/close char (#1)
If the open and close chars are different, we have to count open and close.  This was not
working if the cursor was *on* one of the characters.  If open the open char, the opening char
was "found" and increased the level to 2 immediately.  When on the close, something similar
happened.

Also converted a cond with only two branches to an if.  (It probably started out with more and
I whittled it down but didn't notice it could be simplified.)

Fixes #1