Fix: Handle empty lines and errors gracefully during insert
This factors out determining the bounds to new helper function
`surround--infer-bounds` during `surround-insert`.

The new function handles a third case: When the line is empty, insert
the pair at point.

The new function has an optional parameter that signals an error when
the bounds couldn't be inferred. `surround-insert` sets this to t to
avoid potentially passing `nil` to `goto-char` which itself signals an
error that's not easy to reason about (`number-or-marker-p`).